+++ mmids-gstreamer/src/endpoints/mod.rs
@@ -1,1 +1,3 @@
+//! Endpoints that interact with gstreamer.
+
 pub mod gst_transcoder;
