+++ mmids-gstreamer/src/lib.rs
@@ -1,3 +1,6 @@
+//! This crate contains all components for mmids applications to interact with with gstreamer
+//! pipelines.
+
 pub mod encoders;
 pub mod endpoints;
 pub mod steps;
